ref(events): Split Slack webhook event handling
Move Slack-specific webhook body handling into a route submodule and keep the generic webhook handler focused on platform dispatch. Remove unsupported event prompt policy fields from the frontmatter schema so future behavior is not reserved without executor enforcement.

Refs #435
